refactor: 도메인 패키지를 기준으로 테스트 코드 패키지 이동
